component separation and profile.vue
The message box component app bar has been separated into its own component and all of the menus in the app bar are now separated as components as well. Also removed the views subfolder so that all views are in the same /views folder and reconfigured their paths in the router. Profile.vue has been added and laid out but no content has been added.