DiscoverDrawer.vue  and DiscoverAppBar.vue
The discover drawer now applies an overlay effect and overlaps the main content when opened similarly to how a temporary drawer would but the rail state is maintained when the drawer is closed. Some modifications to the main layoud depending on screen width have been added to achieve this effect. DiscoverAppBar.vue that displays search queries has been added.