BookShelf.vue and styling changes
BookShelf.vue uses book card components to display books that the user has saved, liked, and commented on(none of the functionality of these has been implemented yet). Styling changes have been made to the book cards so they are larger and have a list of buttons for various functions. Styling changes have been made to the default layout dynamic padding so that it is only added to the discover route. Signup.vue now returns errors from the backend and displays them in below the inputs.