Review.vue and Reviews.vue
Added a sidebar the the book card component with a list of buttons with various functions. The review button has been configured so that you can view the current reviews list and write a review of your own. The like button has toggling functionality but none of the other buttons intended purposes are currently configured.