books query and style changes
Users can now query for books from the google books api in the discover page using the search bar or the list of predefined genres in the discover drawer. Styling changes have been made to the book cards to allow the cover images to scale properly and the list of books is no longer displayed in columns but instead they have width thresholds and will wrap to the next line when that threshold is met. Error messages are now displayed in the login page as well.