sending messages
Text messages can now be sent through the message bar in the messages app and book messages can be sent from the discover page or from your bookshelf. ThreadSelecter.vue has been added as a menu so that users can select the thread they wish to send a book message to.